requirements.txt
firebase-admin>=6.2.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
flask>=2.3.3 